Remove native flag and add kernel,stdlib to app deps
With Riak, we ship an erlang VM without HiPE which causes
warnings when bear tries to use the `native` flag to compile.

Also, this commit adds [kernel,stdlib] to bear's application
list in the .app file.  This hopefully will help with shutdown
issues in riak when applications aren't unloaded properly.